Move pass to well_formed, check for groups
