+++ tests/errors/comb-component-groups.expect
@@ -1,0 +1,4 @@
+---CODE---
+1
+---STDERR---
+Error: Malformed Structure: Component `custom_lt` is marked combinational but contains a group.
